Refactor data files
