+++ app/[locale]/about/page.tsx
@@ -1,6 +1,6 @@
 import { Link } from "@/app/navigation";
 import Mail from "@/app/icons/Mail.svg?react";
-import data from "@/data/data.json";
+import data from "@/data/data/cafes.json";
 import BasePage from "@/app/components/BasePage";
 import { Metadata } from "next";
 import { BASE_URL } from "@/app/constants";
