+++ data/loadJSON.mjs
@@ -2,8 +2,8 @@ import path from "path";
 import { promises as fs } from "fs";
 import { dataFolder } from "./constants.mjs";
 
-export default async function saveJSON(fileName, data) {
+export default async function loadJSON(fileName) {
   const filePath = path.join(dataFolder, fileName);
-  const jsonData = JSON.stringify(data, null, 2);
-  await fs.writeFile(filePath, jsonData, "utf8");
+  const rawData = await fs.readFile(filePath, "utf8");
+  return JSON.parse(rawData);
 }
