+++ package.json
@@ -8,7 +8,7 @@
     "build:analyze": "npm run build --analyzeBundle",
     "start": "next start",
     "lint": "next lint",
-    "update-data": "node data/update-data.mjs",
+    "update-cafes": "node data/update-cafes.mjs",
     "update-map-data": "node data/update-map-data.mjs",
     "update-holidays-data": "node data/update-holidays-data.mjs",
     "postinstall": "patch-package",
